+++ src/Extracting/Strategies/Headers/GetFromRouteRules.php
@@ -7,7 +7,7 @@ use Knuckles\Scribe\Extracting\Strategies\Strategy;
 
 class GetFromRouteRules extends Strategy
 {
-    public function __invoke(ExtractedEndpointData $endpointData, array $routeRules): array
+    public function __invoke(ExtractedEndpointData $endpointData, array $routeRules = []): array
     {
         return $routeRules['headers'] ?? [];
     }