+++ src/Extracting/Strategies/Metadata/GetFromDocBlocks.php
@@ -9,7 +9,7 @@ use Mpociot\Reflection\DocBlock;
 
 class GetFromDocBlocks extends Strategy
 {
-    public function __invoke(ExtractedEndpointData $endpointData, array $routeRules): array
+    public function __invoke(ExtractedEndpointData $endpointData, array $routeRules = []): array
     {
         $docBlocks = RouteDocBlocker::getDocBlocksFromRoute($endpointData->route);
         $methodDocBlock = $docBlocks['method'];