+++ src/Extracting/Strategies/Responses/UseApiResourceTags.php
@@ -30,7 +30,7 @@ class UseApiResourceTags extends Strategy
     use DatabaseTransactionHelpers;
     use InstantiatesExampleModels;
 
-    public function __invoke(ExtractedEndpointData $endpointData, array $routeRules): ?array
+    public function __invoke(ExtractedEndpointData $endpointData, array $routeRules = []): ?array
     {
         $methodDocBlock = RouteDocBlocker::getDocBlocksFromRoute($endpointData->route)['method'];