+++ src/Extracting/Strategies/Responses/UseTransformerTags.php
@@ -28,7 +28,7 @@ class UseTransformerTags extends Strategy
 {
     use DatabaseTransactionHelpers, InstantiatesExampleModels;
 
-    public function __invoke(ExtractedEndpointData $endpointData, array $routeRules): ?array
+    public function __invoke(ExtractedEndpointData $endpointData, array $routeRules = []): ?array
     {
         $docBlocks = RouteDocBlocker::getDocBlocksFromRoute($endpointData->route);
         $methodDocBlock = $docBlocks['method'];