+++ tests/Fixtures/TestCustomEndpointMetadata.php
@@ -7,7 +7,7 @@ use Knuckles\Scribe\Extracting\Strategies\Strategy;
 
 class TestCustomEndpointMetadata extends Strategy
 {
-    public function __invoke(ExtractedEndpointData $endpointData, array $routeRules): ?array
+    public function __invoke(ExtractedEndpointData $endpointData, array $routeRules = []): ?array
     {
         $endpointData->metadata->custom['myProperty'] = 'some custom metadata';
         return null;